refactor(opencode): rewrite cli export to use Export.session core
